requirements.txt
ultralytics
PyQt5
RPi.GPIO